Use appropriate error for incorrect reserved bits in TPMA structures (#39)
According to the TPM 2.0 Library specification, Part 2 (rev. 1.38),
Table 2, a TPM_RC_RESERVED_BITS error shall be used if "a non-zero value
was found in a reserved field of an attribute structure (TPMA_)". This
is done correctly when compiling with TABLE_DRIVEN_MARSHAL=NO, cf.
Marshal.c, but for TABLE_DRIVEN_MARSHAL=YES the incorrect error
TPM_RC_ATTRIBUTES is used instead.

Fixes: f76525015bd33fb83933f958d712cf975a5e30ae