Unconditionally run fseek in NvFileSize. (#49)
The line

assert(fseek(s_NvFile, 0, SEEK_END) == 0);

is not run in NDEBUG mode, due to asserts being removed. However, fseek
still has to be run for ftell to properly report the complete file size
in NvFileSize. This separates out the fseek call so it works in NDEBUG
and DEBUG modes.