Fix compiler warnings which may be treated as errors
CompilerDependencies.h: __pragma is commonly defined in RTOS used as
part of various firmware platforms. Only define it if it isn't already
defined.

TpmSizeChecks.c: `once` is always defined, but is only used #if DEBUG.
This can cause "defined but not used" compiler errors.

Global.h and SessionProcess.c: Some areas incorrectly check for
`#ifdef TPM_CC_GetCommandAuditDigest` (which is always defined)
instead of checking for `#if CC_GetCommandAuditDigest`. This leads
to various "defined but not used" errors.